+++ changelog.md
@@ -786,3 +786,4 @@
 #### 7 Aug. 2023
 
 - Fixed a bug where dashing into the boundary between the boundary between two directly adjacent block causes the player to get stuck
+- Fixed a bug where you can add/rename saves/rooms with an empty name